refactor: Move integration interfaces into new core package
