+++ internal/core/destination.go
@@ -1,4 +1,4 @@
-package destination
+package core
 
 type Destination interface {
 	UpdateUsers(users []string) error
