+++ internal/core/source.go
@@ -1,4 +1,4 @@
-package source
+package core
 
 type Source interface {
 	FetchUsers() ([]string, error)
