+++ internal/integration/example_destination.go
@@ -4,7 +4,7 @@ import (
 	"github.com/go-viper/mapstructure/v2"
 
 	"github.com/kiyanmair/shift-sync/config"
-	"github.com/kiyanmair/shift-sync/internal/destination"
+	"github.com/kiyanmair/shift-sync/internal/core"
 	"github.com/kiyanmair/shift-sync/internal/registry"
 )
 
@@ -18,7 +18,7 @@ type ExampleDestination struct {
 	GroupID string `mapstructure:"group_id"`
 }
 
-func NewExampleDestination(cfg config.Destination) (destination.Destination, error) {
+func NewExampleDestination(cfg config.Destination) (core.Destination, error) {
 	var dest ExampleDestination
 	dest.ID = cfg.ID
