+++ internal/registry/source.go
@@ -4,16 +4,16 @@ import (
 	"fmt"
 
 	"github.com/kiyanmair/shift-sync/config"
-	"github.com/kiyanmair/shift-sync/internal/source"
+	"github.com/kiyanmair/shift-sync/internal/core"
 )
 
-var sourceRegistry = map[string]func(config.Source) (source.Source, error){}
+var sourceRegistry = map[string]func(config.Source) (core.Source, error){}
 
-func RegisterSource(name string, constructor func(config.Source) (source.Source, error)) {
+func RegisterSource(name string, constructor func(config.Source) (core.Source, error)) {
 	sourceRegistry[name] = constructor
 }
 
-func NewSource(config config.Source) (source.Source, error) {
+func NewSource(config config.Source) (core.Source, error) {
 	constructor, exists := sourceRegistry[config.Type]
 	if !exists {
 		return nil, fmt.Errorf("unsupported source type: %s", config.Type)
