+++ internal/sync/syncer.go
@@ -4,15 +4,14 @@ import (
 	"log"
 
 	"github.com/kiyanmair/shift-sync/config"
-	"github.com/kiyanmair/shift-sync/internal/destination"
+	"github.com/kiyanmair/shift-sync/internal/core"
 	"github.com/kiyanmair/shift-sync/internal/registry"
-	"github.com/kiyanmair/shift-sync/internal/source"
 )
 
 type Syncer struct {
 	cfg          *config.Config
-	sources      map[string]source.Source
-	destinations map[string]destination.Destination
+	sources      map[string]core.Source
+	destinations map[string]core.Destination
 }
 
 func NewSyncer(configPath string) *Syncer {
@@ -21,7 +20,7 @@ func NewSyncer(configPath string) *Syncer {
 		log.Fatalf("Failed to load config: %v", err)
 	}
 
-	sources := make(map[string]source.Source)
+	sources := make(map[string]core.Source)
 	for _, srcCfg := range cfg.Sources {
 		src, err := registry.NewSource(srcCfg)
 		if err != nil {
@@ -31,7 +30,7 @@ func NewSyncer(configPath string) *Syncer {
 		sources[srcCfg.ID] = src
 	}
 
-	destinations := make(map[string]destination.Destination)
+	destinations := make(map[string]core.Destination)
 	for _, destCfg := range cfg.Destinations {
 		dest, err := registry.NewDestination(destCfg)
 		if err != nil {
